chore: rename from query to queries
