+++ src/pages/[slug]/index.tsx
@@ -5,8 +5,7 @@ import { generateSSGHelper } from "~/server/helper/ssgHelper";
 import { Feed, UserLayout } from "~/components/layouts";
 import UserNotFound from "~/components/user-not-found";
 import { useInView } from "react-intersection-observer";
-import { getUserbyUsername } from "~/hooks/query";
-import { SEO } from "~/components/simple-seo";
+import { getUserbyUsername } from "~/hooks/queries";
 
 const ProfilePage: NextPage<{ username: string }> = ({ username }) => {
   const { data: user } = getUserbyUsername({ username });
